+++ BraintreeUIKit/Components/BTUIKExpiryInputView.m
@@ -3,6 +3,7 @@
 #import "BTUIKCollectionReusableView.h"
 #import "BTUIKAppearance.h"
 #import "BTUIKLocalizedString.h"
+#import "BTUIKViewUtil.h"
 
 #define BT_EXPIRY_FULL_PADDING 10
 #define BT_EXPIRY_SECTION_HEADER_HEIGHT 12
@@ -284,7 +285,7 @@ referenceSizeForHeaderInSection:(__unused NSInteger)section {
 - (CGSize)collectionView:(UICollectionView *)collectionView
                   layout:(__unused UICollectionViewLayout *)collectionViewLayout
   sizeForItemAtIndexPath:(__unused NSIndexPath *)indexPath {
-    BOOL isLandscape = UIInterfaceOrientationIsLandscape(UIApplication.sharedApplication.statusBarOrientation);
+    BOOL isLandscape = [BTUIKViewUtil isOrientationLandscape];
     int monthRows = isLandscape ? 3.0 : 4.0;
     
     CGFloat cellHeight = (CGRectGetHeight(self.monthCollectionView.frame) - BT_EXPIRY_SECTION_HEADER_HEIGHT - ((BT_EXPIRY_FULL_PADDING * monthRows) + BT_EXPIRY_FULL_PADDING)) / monthRows;