+++ BraintreeUIKit/Public/BTUIKViewUtil.h
@@ -84,4 +84,13 @@ typedef NS_ENUM(NSInteger, BTUIKVectorArtSize) {
 + (NSTextAlignment)naturalTextAlignment;
 /// @return NSTextAlignmentLeft if isLanguageLayoutDirectionRightToLeft is true. Ohterwise NSTextAlignmentRight.
 + (NSTextAlignment)naturalTextAlignmentInverse;
+
+#pragma mark Orientation Utilities
+
+/// @return true if the device is in landscape
++ (BOOL)isOrientationLandscape;
+
+/// @return height of the status bar
++ (CGFloat)statusBarHeight;
+
 @end